Fix tests

+++ tests/pipeline/preprocessing/test-register_pairs_handling.py
@@ -155,7 +155,7 @@ def test_undefined_register_pair_is_replaced():
             replacement_var := Variable("loc_0", Integer.uint64_t(), 0),
             operations.BinaryOperation(
                 operations.OperationType.plus,
-                [eax, operations.BinaryOperation(operations.OperationType.left_shift, [ebx, Constant(eax.type.size, Integer.int32_t())])],
+                [eax, operations.BinaryOperation(operations.OperationType.left_shift, [ebx, Constant(eax.type.size, Integer.uint8_t())])],
             ),
         ),
         Assignment(operations.ListOperation([]), operations.Call(FunctionSymbol("kill", 0), [replacement_var])),
@@ -213,7 +213,7 @@ def test_definition_is_placed_before_usage():
             v(),
             BinaryOperation(
                 OperationType.plus,
-                [eax(), BinaryOperation(OperationType.left_shift, [ebx(), Constant(eax().type.size, Integer.int32_t())])],
+                [eax(), BinaryOperation(OperationType.left_shift, [ebx(), Constant(eax().type.size, Integer.uint8_t())])],
             ),
         ),
         Assignment(ListOperation([]), Call(bar(), [v()])),
@@ -312,7 +312,7 @@ def test_definition_is_inserted_correctly():
         Assignment(ebx.copy(), Call(FunctionSymbol("foo", 0), [eax.copy()])),
         Assignment(
             Variable("loc_0", Integer.int64_t(), 0),
-            BinaryOperation(OperationType.plus, [eax.copy(), BinaryOperation(OperationType.left_shift, [ebx.copy(), Constant(32, Integer.int32_t())])]),
+            BinaryOperation(OperationType.plus, [eax.copy(), BinaryOperation(OperationType.left_shift, [ebx.copy(), Constant(32, Integer.uint8_t())])]),
         ),
         Branch(Condition(OperationType.greater, [eax.copy(), ebx.copy()])),
     ]
